pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["scrapers", "pipelines"]

[project]
name = "tnjindex"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.9"
dependencies = [
    "aiotieba>=4.6.0",
    "beautifulsoup4>=4.14.3",
    "dashscope>=1.14.0",
    "imagehash>=4.3.2",
    "openai>=1.0.0",
    "pillow>=11.3.0",
    "pysqlite3>=0.5.2",
    "python-dotenv>=1.0.0",
    "requests>=2.32.5",
    "sqlite-vec>=0.1.6",
]
